+++ planner/core/plan_cache.go
@@ -27,13 +27,11 @@ import (
 	"github.com/pingcap/tidb/parser/ast"
 	"github.com/pingcap/tidb/parser/mysql"
 	core_metrics "github.com/pingcap/tidb/planner/core/metrics"
-	"github.com/pingcap/tidb/planner/util"
 	"github.com/pingcap/tidb/planner/util/debugtrace"
 	"github.com/pingcap/tidb/privilege"
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/stmtctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
-	"github.com/pingcap/tidb/sessiontxn/staleread"
 	"github.com/pingcap/tidb/table/tables"
 	"github.com/pingcap/tidb/types"
 	driver "github.com/pingcap/tidb/types/parser_driver"
@@ -809,129 +807,4 @@ func GetBindSQL4PlanCache(sctx sessionctx.Context, stmt *PlanCacheStmt) (string,
 		}
 	}
 	return "", ignore
-}
-
-// IsPointPlanShortPathOK check if we can execute using plan cached in prepared structure
-// Be careful with the short path, current precondition is ths cached plan satisfying
-// IsPointGetWithPKOrUniqueKeyByAutoCommit
-func IsPointPlanShortPathOK(sctx sessionctx.Context, is infoschema.InfoSchema, stmt *PlanCacheStmt) (bool, error) {
-	stmtAst := stmt.PreparedAst
-	if stmtAst.CachedPlan == nil || staleread.IsStmtStaleness(sctx) {
-		return false, nil
-	}
-	// check auto commit
-	if !IsAutoCommitTxn(sctx) {
-		return false, nil
-	}
-	if stmtAst.SchemaVersion != is.SchemaMetaVersion() {
-		stmtAst.CachedPlan = nil
-		stmt.ColumnInfos = nil
-		return false, nil
-	}
-	// maybe we'd better check cached plan type here, current
-	// only point select/update will be cached, see "getPhysicalPlan" func
-	var ok bool
-	var err error
-	switch stmtAst.CachedPlan.(type) {
-	case *PointGetPlan:
-		ok = true
-	case *Update:
-		pointUpdate := stmtAst.CachedPlan.(*Update)
-		_, ok = pointUpdate.SelectPlan.(*PointGetPlan)
-		if !ok {
-			err = errors.Errorf("cached update plan not point update")
-			stmtAst.CachedPlan = nil
-			return false, err
-		}
-	default:
-		ok = false
-	}
-	return ok, err
-}
-
-func isSafePointGetPath4PlanCache(sctx sessionctx.Context, path *util.AccessPath) bool {
-	// PointGet might contain some over-optimized assumptions, like `a>=1 and a<=1` --> `a=1`, but
-	// these assumptions may be broken after parameters change.
-
-	if isSafePointGetPath4PlanCacheScenario1(path) {
-		return true
-	}
-
-	// TODO: enable this fix control switch by default after more test cases are added.
-	if sctx != nil && sctx.GetSessionVars() != nil && sctx.GetSessionVars().OptimizerFixControl != nil {
-		v, ok := sctx.GetSessionVars().OptimizerFixControl[variable.TiDBOptFixControl44830]
-		if ok && variable.TiDBOptOn(v) && (isSafePointGetPath4PlanCacheScenario2(path) || isSafePointGetPath4PlanCacheScenario3(path)) {
-			return true
-		}
-	}
-
-	return false
-}
-
-func isSafePointGetPath4PlanCacheScenario1(path *util.AccessPath) bool {
-	// safe scenario 1: each column corresponds to a single EQ, `a=1 and b=2 and c=3` --> `[1, 2, 3]`
-	if len(path.Ranges) <= 0 || path.Ranges[0].Width() != len(path.AccessConds) {
-		return false
-	}
-	for _, accessCond := range path.AccessConds {
-		f, ok := accessCond.(*expression.ScalarFunction)
-		if !ok || f.FuncName.L != ast.EQ { // column = constant
-			return false
-		}
-	}
-	return true
-}
-
-func isSafePointGetPath4PlanCacheScenario2(path *util.AccessPath) bool {
-	// safe scenario 2: this Batch or PointGet is simply from a single IN predicate, `key in (...)`
-	if len(path.Ranges) <= 0 || len(path.AccessConds) != 1 {
-		return false
-	}
-	f, ok := path.AccessConds[0].(*expression.ScalarFunction)
-	if !ok || f.FuncName.L != ast.In {
-		return false
-	}
-	return len(path.Ranges) == len(f.GetArgs())-1 // no duplicated values in this in-list for safety.
-}
-
-func isSafePointGetPath4PlanCacheScenario3(path *util.AccessPath) bool {
-	// safe scenario 3: this Batch or PointGet is simply from a simple DNF like `key=? or key=? or key=?`
-	if len(path.Ranges) <= 0 || len(path.AccessConds) != 1 {
-		return false
-	}
-	f, ok := path.AccessConds[0].(*expression.ScalarFunction)
-	if !ok || f.FuncName.L != ast.LogicOr {
-		return false
-	}
-
-	dnfExprs := expression.FlattenDNFConditions(f)
-	if len(path.Ranges) != len(dnfExprs) {
-		// no duplicated values in this in-list for safety.
-		// e.g. `k=1 or k=2 or k=1` --> [[1, 1], [2, 2]]
-		return false
-	}
-
-	for _, expr := range dnfExprs {
-		f, ok := expr.(*expression.ScalarFunction)
-		if !ok {
-			return false
-		}
-		switch f.FuncName.L {
-		case ast.EQ: // (k=1 or k=2) --> [k=1, k=2]
-		case ast.LogicAnd: // ((k1=1 and k2=1) or (k1=2 and k2=2)) --> [k1=1 and k2=1, k2=2 and k2=2]
-			cnfExprs := expression.FlattenCNFConditions(f)
-			if path.Ranges[0].Width() != len(cnfExprs) { // not all key columns are specified
-				return false
-			}
-			for _, expr := range cnfExprs { // k1=1 and k2=1
-				f, ok := expr.(*expression.ScalarFunction)
-				if !ok || f.FuncName.L != ast.EQ {
-					return false
-				}
-			}
-		default:
-			return false
-		}
-	}
-	return true
 }
